Kafka Consumer

+++ src/main/java/com/consumer/kafkaConsumer/repository/LibraryEventRepository.java
@@ -2,7 +2,10 @@ package com.consumer.kafkaConsumer.repository;
 
 
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +13,8 @@ import com.consumer.kafkaConsumer.entity.LibraryEvent;
 
 @Repository
 @Component
-public interface LibraryEventRepository extends  JpaRepository<LibraryEvent, Long> {
+public interface LibraryEventRepository extends  JpaRepository<LibraryEvent, Integer> {
+
+	 
 
 }
